Add files via upload
This mod changes the way critical damage is dealt. Instead of a blanket 2X modifier to the base damage of a weapon before it's calculated by any other source, It's a different modifier depending on the manufacturer. 

Jakobs now get a 50% boost in damage for crits, Dahl 30%, Hyperion 35%, Vladof 20%, COV 20%, Maliwan 15%, Atlas 20%, and Torgue is unchanged as of now. 

The base damage of each of these weapons had been brought up to the point where the final damage with their new critical modifier is EXACTLY THE SAME as it was when the modifier was 2X. This leads to a high body shot damage for each weapon. After the change, Dahl was given a 4% base damage nerf, Hyperion a 5% damage nerf, 6% damage perf for COV and Valdof, and a 7% nerf for Mailwan, all to compensate for the skill reduction inherent in this new system. Atlas weapons have their own damage changes adjusted to incentivize users to use smart bullets. No need to worry about your puck hitting a non critical location on an enemy, which further incentivizes using the tracking mechanic.  

The base strength of each of these weapons is based on the EpicNNG's Borderlands 3 Redux mod. https://github.com/BLCM/bl3mods/wiki/Redux%20Mod

This is my first mod! And I would appreciate feedback.